refactor(dnd): extract useTabDragAndDrop hook and TabDragOverlay component
Split WindowGroupList (435 lines) into:
- src/hooks/useTabDragAndDrop.tsx — state, sensors, collision detection,
  cross-window ring highlight effect, and all drag handlers. Uses ref
  mirror pattern (allTabsMap / filteredTabGroups / dropPosition) so
  useCallback([]) handlers stay stable while reading fresh state, preventing
  stale-closure bugs on mid-drag UPDATE_TABS from background.
- src/components/TabDragOverlay.tsx — drag preview clone and multi-drag
  badge (consumes DragSelectionContext internally).

WindowGroupList is now 64 lines: props definition, hook consumption, and
DndContext / DragOverlay wrappers.

Additional cleanups (Section 3 leftover):
- allTabs is now a Map<number, Tab> for O(1) lookup on every handler.
- Cross-window branch's two allTabs.find(...)! non-null assertions merged
  into a single lookup with early return + toast ("Drag target went
  missing, please retry") on miss.
- KeyboardSensor + sortableKeyboardCoordinates moved into the hook,
  behavior unchanged.

Non-asymmetric clearDragSelection semantics preserved:
- same-window: only clear when the dragged tab was NOT part of the
  drag-selection set.
- cross-window: always clear.

Testing:
- E2E: new e2e/keyboard-drag.spec.ts covers Enter → ArrowDown → Enter
  reorder and Escape cancel. cross-window-dnd (11) and multi-drag (6)
  continue to pass unchanged.

Plan: ~/.claude/plans/lazycluster-section6-windowgrouplist-split.md PR 2.
