+++ e2e/keyboard-drag.spec.ts
@@ -1,0 +1,73 @@
+import { test, expect } from './fixtures';
+import { createTabs } from './helpers';
+
+// Read tab ids in current-window order via Chrome API — resilient to
+// duplicate titles/hrefs across pre-existing `about:blank` fixture tabs.
+async function readTabIds(page: import('@playwright/test').Page): Promise<number[]> {
+  return page.evaluate(async () => {
+    const win = await chrome.windows.getCurrent();
+    const tabs = await chrome.tabs.query({ windowId: win.id });
+    return tabs.map(t => t.id!);
+  });
+}
+
+test.describe('Keyboard Drag E2E Tests', () => {
+  test('should reorder tab via keyboard drag (Enter → ArrowDown → Enter)', async ({ page, extensionId }) => {
+    await page.goto(`chrome-extension://${extensionId}/manager.html`);
+    await page.locator('.group\\/tabitem').first().waitFor({ state: 'visible' });
+
+    const initialCount = await page.locator('.group\\/tabitem').count();
+    if (initialCount < 3) {
+      await createTabs(page, 3 - initialCount);
+      await page.reload();
+      await page.locator('.group\\/tabitem').first().waitFor({ state: 'visible' });
+    }
+
+    const tabIdsBefore = await readTabIds(page);
+    expect(tabIdsBefore.length).toBeGreaterThanOrEqual(3);
+    const originalFirstId = tabIdsBefore[0];
+
+    // Activate keyboard drag on first tab's handle
+    const firstDragHandle = page.locator('button[aria-label="Drag to reorder"]').first();
+    await firstDragHandle.focus();
+    await page.keyboard.press('Enter'); // Start drag
+    await page.keyboard.press('ArrowDown'); // Move over tab 1
+    await page.keyboard.press('ArrowDown'); // Move over tab 2
+    await page.keyboard.press('Enter'); // Drop
+
+    // Wait for the Chrome API update to propagate: first tab id changed
+    await expect.poll(async () => (await readTabIds(page))[0]).not.toBe(originalFirstId);
+
+    const tabIdsAfter = await readTabIds(page);
+    // Original first tab moved later in the window
+    const newIndex = tabIdsAfter.indexOf(originalFirstId);
+    expect(newIndex).toBeGreaterThan(0);
+  });
+
+  test('should cancel keyboard drag on Escape and keep original order', async ({ page, extensionId }) => {
+    await page.goto(`chrome-extension://${extensionId}/manager.html`);
+    await page.locator('.group\\/tabitem').first().waitFor({ state: 'visible' });
+
+    const initialCount = await page.locator('.group\\/tabitem').count();
+    if (initialCount < 3) {
+      await createTabs(page, 3 - initialCount);
+      await page.reload();
+      await page.locator('.group\\/tabitem').first().waitFor({ state: 'visible' });
+    }
+
+    const tabIdsBefore = await readTabIds(page);
+    expect(tabIdsBefore.length).toBeGreaterThanOrEqual(3);
+
+    // Start keyboard drag, move once, then cancel with Escape
+    const firstDragHandle = page.locator('button[aria-label="Drag to reorder"]').first();
+    await firstDragHandle.focus();
+    await page.keyboard.press('Enter'); // Start drag
+    await page.keyboard.press('ArrowDown'); // Move (preview only)
+    await page.keyboard.press('Escape'); // Cancel — should trigger onDragCancel
+
+    // Assert the pre-existing tab order is unchanged. Slice off any tabs
+    // that appear during the poll window (test environment may spawn stray
+    // tabs on some runs), which is orthogonal to whether Escape cancelled.
+    await expect.poll(async () => (await readTabIds(page)).slice(0, tabIdsBefore.length)).toEqual(tabIdsBefore);
+  });
+});
