+++ src/components/WindowGroupList.tsx
@@ -1,391 +1,37 @@
-import { useState, useEffect, useRef, useMemo } from 'react';
-import {
-  DndContext,
-  DragEndEvent,
-  DragOverEvent,
-  DragOverlay,
-  DragStartEvent,
-  PointerSensor,
-  KeyboardSensor,
-  useSensor,
-  useSensors,
-  pointerWithin,
-  rectIntersection,
-  CollisionDetection,
-} from '@dnd-kit/core';
-import { sortableKeyboardCoordinates } from '@dnd-kit/sortable';
+import { DndContext, DragOverlay } from '@dnd-kit/core';
 import WindowGroup from './WindowGroup';
-import TabItem from './TabItem';
+import TabDragOverlay from './TabDragOverlay';
 import { WindowGroupContextProvider } from '../contexts/WindowGroupContext';
-import { useToast } from './ToastProvider';
-import Alert from './Alert';
-import { useDragSelectionContext } from '../contexts/DragSelectionContext';
-import { identifyGroupsToPreserve } from '../utils/tabGroupPreservation';
-import { buildMoveSegments } from '../utils/tabMove';
-import { getSortableTabData, getWindowGroupDropData } from '../types/dnd';
+import { useTabDragAndDrop, type FilteredTabGroup } from '../hooks/useTabDragAndDrop';
 
 interface WindowGroupListProps {
-  filteredTabGroups: { windowId: number; tabs: chrome.tabs.Tab[]; windowGroupNumber: number }[];
+  filteredTabGroups: FilteredTabGroup[];
   isFiltered?: boolean;
 }
 
 const WindowGroupList = ({ filteredTabGroups, isFiltered = false }: WindowGroupListProps) => {
-  const { showToast } = useToast();
-  const { clearDragSelection, dragSelectedTabIds } = useDragSelectionContext();
-
-  // Track active dragging item for DragOverlay
-  const [activeId, setActiveId] = useState<number | null>(null);
-
-  // Track drop indicator position
-  const [overId, setOverId] = useState<number | null>(null);
-  const [dropPosition, setDropPosition] = useState<'top' | 'bottom'>('bottom');
-
-  // Track which window group is being dragged over (for cross-window ring highlight)
-  const [overWindowId, setOverWindowId] = useState<number | null>(null);
-
-  // Configure sensors for drag-and-drop
-  const sensors = useSensors(
-    useSensor(PointerSensor, {
-      activationConstraint: { distance: 8 }, // Prevent accidental drags
-    }),
-    useSensor(KeyboardSensor, {
-      coordinateGetter: sortableKeyboardCoordinates,
-      keyboardCodes: {
-        start: ['Enter'], // Only Enter key activates drag (Space is reserved for checkbox toggle)
-        cancel: ['Escape'],
-        end: ['Enter'],
-      },
-    })
-  );
-
-  // Pointer-first collision detection for multi-container DnD
-  const collisionDetection: CollisionDetection = args => {
-    const pointerCollisions = pointerWithin(args);
-    if (pointerCollisions.length > 0) return pointerCollisions;
-    return rectIntersection(args);
-  };
-
-  // All tabs across all windows (for lookups); memoized so drag-state renders
-  // don't rebuild it
-  const allTabs = useMemo(() => filteredTabGroups.flatMap(g => g.tabs), [filteredTabGroups]);
-
-  // Store source window ID for pointer tracking (set in handleDragStart, read in useEffect)
-  const sourceWindowIdRef = useRef<number | null>(null);
-
-  // Track pointer position during drag for reliable ring highlight.
-  // Uses DOM elementsFromPoint() instead of dnd-kit collision detection to avoid
-  // flickering caused by rectIntersection fallback matching adjacent window groups.
-  useEffect(() => {
-    if (activeId === null) return;
-
-    const sourceWindowId = sourceWindowIdRef.current;
-
-    const handlePointerMove = (e: PointerEvent) => {
-      const elements = document.elementsFromPoint(e.clientX, e.clientY);
-      const windowGroupEl = elements.find(
-        (el): el is HTMLElement => el instanceof HTMLElement && 'windowId' in el.dataset
-      );
-      if (windowGroupEl) {
-        const targetWindowId = Number(windowGroupEl.dataset.windowId);
-        setOverWindowId(targetWindowId !== sourceWindowId ? targetWindowId : null);
-      } else {
-        setOverWindowId(null);
-      }
-    };
-
-    document.addEventListener('pointermove', handlePointerMove);
-    return () => {
-      document.removeEventListener('pointermove', handlePointerMove);
-      setOverWindowId(null);
-    };
-  }, [activeId]);
-
-  // Handle drag start to track active item for DragOverlay
-  const handleDragStart = (event: DragStartEvent) => {
-    const newActiveId = Number(event.active.id);
-    const activeTab = allTabs.find(t => t.id === newActiveId);
-    sourceWindowIdRef.current = activeTab?.windowId ?? null;
-    setActiveId(newActiveId);
-  };
-
-  // Handle drag over event to show drop indicator
-  const handleDragOver = (event: DragOverEvent) => {
-    const { active, over } = event;
-
-    if (over) {
-      // Window group droppable (collapsed group or empty area) — no tab-level indicator
-      if (getWindowGroupDropData(over)) {
-        setOverId(null);
-        return;
-      }
-
-      const activeTab = allTabs.find(t => t.id === active.id);
-      const overTab = allTabs.find(t => t.id === over.id);
-
-      if (activeTab && overTab) {
-        setOverId(Number(over.id));
-
-        if (activeTab.windowId === overTab.windowId) {
-          // Same window: direction based on index comparison
-          const position = activeTab.index > overTab.index ? 'top' : 'bottom';
-          setDropPosition(position);
-        } else {
-          // Cross-window: always insert after the over tab
-          setDropPosition('bottom');
-        }
-      } else {
-        setOverId(null);
-      }
-    }
-  };
-
-  // Handle drag end event
-  const handleDragEnd = async (event: DragEndEvent) => {
-    const { active, over } = event;
-
-    // Reset drag overlay and drop indicator
-    setActiveId(null);
-    setOverId(null);
-
-    // Early return: no drop target or dropped on itself
-    if (!over || active.id === over.id) return;
-
-    const activeTab = allTabs.find(t => t.id === active.id);
-    if (!activeTab) return;
-
-    // Determine target window ID based on drop target type
-    const windowGroupDrop = getWindowGroupDropData(over);
-    let targetWindowId: number;
-
-    if (windowGroupDrop) {
-      // Dropped on a window group container (e.g., collapsed group header)
-      targetWindowId = windowGroupDrop.windowId;
-    } else {
-      // Dropped on a specific tab
-      const overTab = allTabs.find(t => t.id === over.id);
-      if (!overTab) {
-        console.error('Drop target tab not found:', over.id);
-        return;
-      }
-      targetWindowId = overTab.windowId!;
-    }
-
-    const isCrossWindow = activeTab.windowId !== targetWindowId;
-
-    if (!isCrossWindow) {
-      // ---- Same-window move (existing logic) ----
-      const overTab = allTabs.find(t => t.id === over.id);
-      if (!overTab) return;
-
-      const windowTabs = filteredTabGroups.find(g => g.windowId === activeTab.windowId)?.tabs ?? [];
-
-      try {
-        // Get drag selection data from useSortable
-        const sortableData = getSortableTabData(active);
-        const selectedItems = sortableData?.selectedItems;
-        const isSelected = sortableData?.isSelected ?? false;
-
-        // Determine which tabs to move
-        let tabsToMove: number[];
-
-        if (isSelected && selectedItems) {
-          // Sort selected tab IDs by their current browser position to preserve original order
-          tabsToMove = selectedItems
-            .map(id => {
-              const tab = windowTabs.find(t => t.id === id);
-              return tab ? { id, index: tab.index } : null;
-            })
-            .filter((item): item is { id: number; index: number } => item !== null)
-            .sort((a, b) => a.index - b.index)
-            .map(item => item.id);
-        } else {
-          tabsToMove = [Number(active.id)];
-        }
-
-        // Identify tab groups to preserve (all members of a group are being moved)
-        const groupsToPreserve = identifyGroupsToPreserve(tabsToMove, windowTabs);
-
-        if (groupsToPreserve.length > 0) {
-          // Segment-based move: use chrome.tabGroups.move() for groups to avoid
-          // chrome.tabs.move() breaking group membership.
-          const preservedGroupIds = new Set(groupsToPreserve.map(g => g.groupId));
-          const tabMap = new Map(windowTabs.map(t => [t.id!, t]));
-          const segments = buildMoveSegments(tabsToMove, tabMap, preservedGroupIds);
-
-          // Step 1: Collect all segments at end of window (preserves relative order)
-          for (const seg of segments) {
-            if (seg.type === 'group') {
-              await chrome.tabGroups.move(seg.groupId, { index: -1 });
-            } else {
-              await chrome.tabs.move(seg.tabIds, { index: -1 });
-            }
-          }
-
-          // Step 2: Get drop target's current index (shifted after collection)
-          const overTabNow = await chrome.tabs.get(overTab.id!);
-          let insertAt = overTabNow.index;
-          if (dropPosition === 'bottom') insertAt++;
-
-          // Step 3: Move segments from end to target position in order
-          let offset = 0;
-          for (const seg of segments) {
-            if (seg.type === 'group') {
-              await chrome.tabGroups.move(seg.groupId, { index: insertAt + offset });
-            } else {
-              await chrome.tabs.move(seg.tabIds, { index: insertAt + offset });
-            }
-            offset += seg.tabIds.length;
-          }
-        } else {
-          // No groups to preserve — sequential move with direction-aware index
-          const selectedIndices = tabsToMove
-            .map(id => windowTabs.find(t => t.id === id)?.index)
-            .filter((index): index is number => index !== undefined);
-          const minSelectedIndex = Math.min(...selectedIndices);
-
-          let targetIndex = overTab.index;
-          if (dropPosition === 'bottom') targetIndex++;
-
-          const isMovingDown = minSelectedIndex < targetIndex;
-          if (isMovingDown) targetIndex -= tabsToMove.length;
-
-          if (tabsToMove.length === 1) {
-            await chrome.tabs.move(tabsToMove[0], { index: targetIndex });
-          } else {
-            if (isMovingDown) {
-              // Downward: move in reverse order (last tab first)
-              for (let i = tabsToMove.length - 1; i >= 0; i--) {
-                await chrome.tabs.move(tabsToMove[i], { index: targetIndex + i });
-              }
-            } else {
-              // Upward: move in forward order (first tab first)
-              for (let i = 0; i < tabsToMove.length; i++) {
-                await chrome.tabs.move(tabsToMove[i], { index: targetIndex + i });
-              }
-            }
-          }
-        }
-
-        // Clear drag selection only if dragged tab was NOT selected
-        if (!isSelected) {
-          clearDragSelection();
-        }
-      } catch (error) {
-        showToast(<Alert message="Failed to move tab" variant="error" />);
-        console.error('Error moving tab:', error);
-      }
-    } else {
-      // ---- Cross-window move ----
-
-      // Get drag selection data from useSortable
-      const sortableData = getSortableTabData(active);
-      const selectedItems = sortableData?.selectedItems;
-      const isSelected = sortableData?.isSelected ?? false;
-
-      // Calculate target index
-      let targetIndex: number;
-      if (windowGroupDrop) {
-        // Dropped on window group container (collapsed group) — append to end
-        targetIndex = -1;
-      } else {
-        // Dropped on a specific tab in the target window
-        const overTab = allTabs.find(t => t.id === over.id)!;
-        targetIndex = dropPosition === 'top' ? overTab.index : overTab.index + 1;
-      }
-
-      // Hoisted: needed for both single-tab and multi-tab group preservation check
-      const sourceWindowTabs = filteredTabGroups.find(g => g.windowId === activeTab.windowId)?.tabs ?? [];
-
-      // Determine which tabs to move (sorted by index for relative order)
-      let tabsToMove: number[];
-
-      if (isSelected && selectedItems && selectedItems.length > 1) {
-        // Multi-tab cross-window move
-        tabsToMove = selectedItems
-          .map(id => {
-            const tab = sourceWindowTabs.find(t => t.id === id);
-            return tab ? { id, index: tab.index } : null;
-          })
-          .filter((item): item is { id: number; index: number } => item !== null)
-          .sort((a, b) => a.index - b.index)
-          .map(item => item.id);
-      } else {
-        // Single-tab cross-window move
-        tabsToMove = [Number(active.id)];
-      }
-
-      try {
-        // Identify tab groups to preserve BEFORE moving
-        const groupsToPreserve = identifyGroupsToPreserve(tabsToMove, sourceWindowTabs);
-
-        if (groupsToPreserve.length > 0) {
-          // Segment-based cross-window move: use chrome.tabGroups.move() for groups
-          const preservedGroupIds = new Set(groupsToPreserve.map(g => g.groupId));
-          const tabMap = new Map(sourceWindowTabs.map(t => [t.id!, t]));
-          const segments = buildMoveSegments(tabsToMove, tabMap, preservedGroupIds);
-
-          // Step 1: Move all segments to target window (append to end, preserves relative order)
-          for (const seg of segments) {
-            if (seg.type === 'group') {
-              await chrome.tabGroups.move(seg.groupId, { windowId: targetWindowId, index: -1 });
-            } else {
-              await chrome.tabs.move(seg.tabIds, { windowId: targetWindowId, index: -1 });
-            }
-          }
-
-          // Step 2: Reposition within target window if dropped on a specific tab
-          if (targetIndex !== -1) {
-            const overTab = allTabs.find(t => t.id === over.id)!;
-            const overTabNow = await chrome.tabs.get(overTab.id!);
-            let insertAt = overTabNow.index;
-            if (dropPosition === 'bottom') insertAt++;
-
-            let offset = 0;
-            for (const seg of segments) {
-              if (seg.type === 'group') {
-                await chrome.tabGroups.move(seg.groupId, { index: insertAt + offset });
-              } else {
-                await chrome.tabs.move(seg.tabIds, { index: insertAt + offset });
-              }
-              offset += seg.tabIds.length;
-            }
-          }
-        } else {
-          // No groups to preserve — existing two-step move
-          await chrome.tabs.move(tabsToMove, { windowId: targetWindowId, index: -1 });
-          if (targetIndex !== -1) {
-            await chrome.tabs.move(tabsToMove, { index: targetIndex });
-          }
-        }
-
-        clearDragSelection();
-      } catch (error) {
-        showToast(<Alert message="Failed to move tab to another window" variant="error" />);
-        console.error('Error moving tab(s) to another window:', error);
-      }
-    }
-  };
-
-  // Handle drag cancel event (e.g., ESC key pressed)
-  const handleDragCancel = () => {
-    setActiveId(null);
-    setOverId(null);
-  };
-
-  // Find active tab data for DragOverlay
-  const activeTab = activeId ? allTabs.find(t => t.id === activeId) : null;
-  const activeTabWindowTabs = activeTab
-    ? (filteredTabGroups.find(g => g.windowId === activeTab.windowId)?.tabs ?? [])
-    : [];
+  const {
+    sensors,
+    collisionDetection,
+    overId,
+    overWindowId,
+    dropPosition,
+    onDragStart,
+    onDragOver,
+    onDragEnd,
+    onDragCancel,
+    activeTab,
+    activeTabWindowTabs,
+  } = useTabDragAndDrop({ filteredTabGroups });
 
   return (
     <DndContext
       sensors={sensors}
       collisionDetection={collisionDetection}
-      onDragStart={handleDragStart}
-      onDragOver={handleDragOver}
-      onDragEnd={handleDragEnd}
-      onDragCancel={handleDragCancel}
+      onDragStart={onDragStart}
+      onDragOver={onDragOver}
+      onDragEnd={onDragEnd}
+      onDragCancel={onDragCancel}
     >
       <div className="lg:columns-2 mt-4">
         {filteredTabGroups
@@ -408,25 +54,8 @@ const WindowGroupList = ({ filteredTabGroups, isFiltered = false }: WindowGroupL
           ))}
       </div>
 
-      {/* DragOverlay shows clone of dragged item with optional selection badge */}
       <DragOverlay>
-        {activeTab ? (
-          <div className="relative">
-            <ul className="list shadow-md">
-              <TabItem
-                tab={activeTab}
-                isFiltered={false}
-                index={activeTabWindowTabs.findIndex(t => t.id === activeId)}
-                windowId={activeTab.windowId!}
-                tabs={activeTabWindowTabs}
-              />
-            </ul>
-            {/* Show badge with selection count during multi-drag */}
-            {dragSelectedTabIds.has(activeId!) && dragSelectedTabIds.size > 1 && (
-              <div className="absolute -top-2 -right-2 badge badge-sm badge-accent">{dragSelectedTabIds.size}</div>
-            )}
-          </div>
-        ) : null}
+        {activeTab ? <TabDragOverlay activeTab={activeTab} activeTabWindowTabs={activeTabWindowTabs} /> : null}
       </DragOverlay>
     </DndContext>
   );
